% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{adjl_from_adjmat}
\alias{adjl_from_adjmat}
\title{Transform from an adjmatrix to an adjlist of variables}
\usage{
adjl_from_adjmat(adj_m)
}
\arguments{
\item{adj_mat}{Matrix}

\item{n_fold}{Number of links contractions (folds)
Default: \code{0}}
}
\value{
tree: degree list, connectivity matrix and weights
}
\description{
This is a wonderful description(X)
}

