% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{adjl_from_pi}
\alias{adjl_from_pi}
\title{Build the network from the already processed Progress Index file}
\usage{
adjl_from_pi(fil)
}
\arguments{
\item{fl}{file location}

\item{n_fold}{Number of links contractions (folds)
Default: \code{0}}
}
\value{
tree: degree list, connectivity matrix and weights
}
\description{
This is a wonderful description(X)
}

