% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{adjl_from_trj}
\alias{adjl_from_trj}
\title{Build the network from the trajectory file}
\usage{
adjl_from_trj(trj, distance_method = 5, clu_radius = NULL,
  clu_hardcut = NULL, cores = NULL, mode = NULL, min_span_tree = TRUE)
}
\arguments{
\item{fl}{file location}

\item{n_fold}{Number of links contractions (folds)
Default: \code{0}}
}
\value{
tree: degree list, connectivity matrix and weights
}
\description{
This is a wonderful description(X)
}
\examples{
none

}

