% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_optimization.R
\name{basin_optimization}
\alias{basin_optimization}
\title{Basin optimization of the SBR}
\usage{
basin_optimization(the_sap, basin_optimization_method = NULL,
  how_fine_search = 100, denat_opt = NULL,
  plot_basin_identification = FALSE, nbins_x_min = NULL,
  nbins_x_max = NULL, nbins_y = nbins_x_max, number_of_clusters = NULL,
  force_matching = FALSE, silent = FALSE)
}
\arguments{
\item{the_sap}{Name of the PROGIDX_<...>.dat or REPIX_<...>.dat file or Data frame with three columns containing, in order, the (sorted) progress index,
the relative time indices and the cut function.}

\item{basin_optimization_method}{Method name for the selection of precise basin. This is a costly procedure which is based on an heuristical obj function w
with the following subgroups.
\itemize{
   \item "\code{uniformity}" not available
   \item "\code{MI_barrier_weighting}" not available yet (the description)
}}

\item{how_fine_search}{This variable define the length of the vector of bins in which is splitted the nbins_x_min-nbins_x_max interval. As the algorithm undergo binary 
search this is a fundamental variable to define the sensitivity of the search.}

\item{denat_opt}{This value can be set to \code{"process_subtraction"} or \code{"poly_interpolation"} and it is defining the removal of 
parabolic artifacts in the kinetic trace. Check \code{\link{basins_recognition}} for details.}

\item{plot_basin_identification}{A logical value indicating whether to display the SAPPHIRE plot with the resulting partitions or not. 
Black partitions are the matched ones, blue ones derive only from the dynamic analysis and orange ones only from the kinetic analysis. 
The green curve is the kinetic annotation (black curve) where the parabolic shape has been subtracted, i.e. the actual curve used for 
the peaks identification. Default value is \code{FALSE}.}

\item{nbins_x_min}{Min number of bins on  x-axis of the 2-D histogram. Default to 7. It is the lower end of the search space.}

\item{nbins_x_max}{Max number of bins on  x-axis of the 2-D histogram. Default to sqrt(nrow(the_sap)). It is the upper end of the search space.}

\item{nbins_y}{Number of bins on the y-axis of the 2-D histogram. Default to nbins_x_max. This option is at the moment ininfluential.}

\item{number_of_clusters}{if basin_optimization_method is active accordingly this must be set to integer.}

\item{force_matching}{Please refer to \code{\link{basins_recognition}} for further details about the match option.}

\item{silent}{A logical value indicating whether the function has to remain silent or not. Default value is \code{FALSE}.}
}
\value{
A list containing
      \itemize{
        \item "\code{Optimal_nbins}" The found number of bins.
        \item "\code{bas}" Output of the final basin recognition run.
        \item "\code{ncl}" Number of clusters finally found.
      }
}
\description{
\code{basin_optimization} uses the information provided by the SAPPHIRE basin recognition in order to define the optimal partition of the data (uses in 
    particulare the number of bins as a tuning variable).
}
\details{
For details regarding the SAPPHIRE plot, please refer to the relative publications \url{http://www.nature.com/articles/srep06264}. 
Main documentation of the original campari software \url{http://campari.sourceforge.net/documentation.html}.
}
\examples{

adjl <- mst_from_trj(trj = matrix(rnorm(1000), nrow = 100, ncol = 10))
ret<-gen_progindex(adjl = adjl)
gen_annotation(ret_data = ret, local_cut_width = 10)
\dontrun{
score_sapphire(the_sap = "PROGIDX_000000000001.dat", ann = rnorm(100))
CampaRi::basin_optimization(the_sap = "PROGIDX_000000000001.dat",  
how_fine_search = 10, number_of_clusters = 2, force_matching = T, silent = F)
}

}
