% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campari_run.R
\name{campari}
\alias{campari}
\title{plotting}
\usage{
campari(wd, data_file, base_name, camp_home)
}
\arguments{
\item{fl}{file location}

\item{n_fold}{Number of links contractions (folds)
Default: \code{0}}
}
\value{
tree: degree list, connectivity matrix and weights
}
\description{
plotting from file(X)
}

