% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{contract_mst}
\alias{contract_mst}
\title{Compute the contraction of the extremes of the mst}
\usage{
contract_mst(adjl, n_fold = 0)
}
\arguments{
\item{adjl}{A list of 3 adj lists: degree list, connectivity matrix and weights}

\item{n_fold}{Number of links contractions (folds)
Default: \code{0}}
}
\value{
adjl modified
}
\description{
This is a wonderful description(X)
}

