% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mahalanobis.R
\name{find_mahalanobis}
\alias{find_mahalanobis}
\title{Find the mahalanobis distance for a set of triplets}
\usage{
find_mahalanobis(clu1e1, clu1e2, clu2, must_be_positive = TRUE)
}
\arguments{
\item{clu1e1}{Elements of the first clustering. Needed shape: (features, number of vectors). n_vectors must be > of features.}

\item{clu2}{As above. The number of elements should be identical for clu1e1, clu1e2 and clu2}

\item{must_be_positive}{If FALSE, it will consider also negative values for the Mahalanobis distance (the convergence can be much longer).}

\item{clu1}{Elements of the first clustering. Needed shape: (features, number of vectors). n_vectors must be > of features.}
}
\value{
It will return a matrix n_features*n_features.
}
\description{
\code{find_mahalanobis} uses a group of vectors belonging to a cluster and a second group of vector belonging to another cluster and learns
     a distance matrix which maximise the distance between different clusters and minimize the distance between elements of the same cluster
}
\details{
For more details on the SAPPHIRE plot, please refer to the main documentation of the original 
campari software \url{http://campari.sourceforge.net/documentation.html}.
}
\seealso{
\code{\link{adjl_from_progindex}}, \code{\link{gen_progindex}}, \code{\link{gen_annotation}}.
}
