% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_annotation.R
\name{gen_annotation}
\alias{gen_annotation}
\title{Compute the annotation of the progrex index}
\usage{
gen_annotation(ret_data, local_cut_width = NULL, snap_start = NULL)
}
\arguments{
\item{ret_data}{asdrubali che corrono (X)}

\item{local_cut_width}{local cut(X)
Default: \code{15000}}

\item{snap_start}{Starting snapshot
Default: \code{0}}
}
\value{
ret_data
}
\description{
\code{gen_annotation} is able to
}
\details{
For details, please refer to the main documentation of the original campari software \url{http://campari.sourceforge.net/documentation.html}
}

