% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{gen_progindex}
\alias{gen_progindex}
\title{Compute the new index along with parent and distance-to-parent information}
\usage{
gen_progindex(adjl, snap_start = 0)
}
\arguments{
\item{adjl}{A list of 3 adj lists: degree list, connectivity matrix and weights}

\item{snap_start}{Starting snapshot
Default: \code{0}}
}
\value{
ret_data
}
\description{
This is a wonderful description(X)
}
\examples{
tree <- adjl_from_pi(fil = "data/PROGIDX_000000000001.dat")

ret_dt <- contract_mst(adjl = tree,n_fold = 0)
}

