% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_covariance.R
\name{generate_covariance}
\alias{generate_covariance}
\title{Create a covariance matrix for each snapshots}
\usage{
generate_covariance(trj, window = NULL, overlapping_reduction = NULL)
}
\arguments{
\item{trj}{Input trajectory (variables on the columns and equal-time spaced snpashots on the row). It must be a \code{matrix} or a \code{data.frame} of numeric.}

\item{window}{Number of snapshots taken before and after the considered snapshot.}

\item{overlapping_reduction}{Not yet supported (It will if a great number of snapshots will be considered in this analysis - snapshot selection)}
}
\value{
It will return a modified trajectory matrix.
}
\description{
\code{generate_covariance} creates a sequence of covariance matrices (one for each snapshot) using a sliding window of trajectory (see \code{window}). 
     It is also possible to reduce the dimensionality using an overlapping reduction (see \code{overlapping_reduction})
}
\details{
For more details on the SAPPHIRE plot, please refer to the main documentation of the original 
campari software \url{http://campari.sourceforge.net/documentation.html}.
}
\seealso{
\code{\link{adjl_from_progindex}}, \code{\link{gen_progindex}}, \code{\link{gen_annotation}}.
}
