% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_network.R
\name{generate_network}
\alias{generate_network}
\title{Create a network for each snapshots}
\usage{
generate_network(trj, window = NULL, overlapping_reduction = NULL)
}
\arguments{
\item{trj}{Input trajectory (variables on the columns and equal-time spaced snpashots on the row). It must be a \code{matrix} or a \code{data.frame} of numeric.}

\item{window}{Number of snapshots taken before and after the considered snapshot.}

\item{overlapping_reduction}{}
}
\value{
It will return a modified trajectory matrix.
}
\description{
\code{generate_network} creates a sequence of network (one for each snapshot) using a sliding window of trajectory (see \code{window}). 
     It is also possible to reduce the dimensionality using an overlapping reduction (see \code{overlapping_reduction})
}
\details{
For more details on the SAPPHIRE plot, please refer to the main documentation of the original campari software \url{http://campari.sourceforge.net/documentation.html}.
}
\seealso{
\code{\link{adjl_from_progindex}}, \code{\link{gen_progindex}}, \code{\link{gen_annotation}}.
}
