% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmrq_kfold.R
\name{gmrq.kfold}
\alias{gmrq.kfold}
\title{GMRQ score from McGibbon RT et al. (2015)}
\usage{
gmrq.kfold(traj, gmrq = c(2:6), lag = 1, kfolds = 5,
  clust.method = c("kmeans", "sbr", "nsbr", "tbc", "dbscan"), tm.opt = NULL,
  clust.param = NULL, preproc = NULL, chunks = FALSE, dir.out = FALSE,
  plot = FALSE, return_plot = FALSE, silent = FALSE, ...)
}
\arguments{
\item{traj}{trajectory matrix.}

\item{gmrq}{number of eigenvalues.}

\item{lag}{number of lags in the Markov state model (MSM)}

\item{kfolds}{cross-validation number of folds}

\item{clust.method}{clustering method. Available is kmeans sbr and nsbr}

\item{tm.opt}{definition for matrix creation (C and S)}

\item{clust.param}{the cluster parameters to optimize - it is a matrix with parameters on the columns and 
different set of them in the rows (it loops on the rows)}

\item{preproc}{vars for the campari run}

\item{chunks}{if \code{TRUE} it splits consecutive chunks of data. It is needed for time-series data.}

\item{dir.out}{directory for files}

\item{plot}{if \code{TRUE} plots.}

\item{return_plot}{if \code{TRUE} return the plot.}

\item{silent}{if \code{TRUE} print less strings.}

\item{...}{mainly debugging options}
}
\value{
A list containing tot
}
\description{
Function to calculate the GMRQ score. Input are the multidim trajectory,vector of GMRQ values, an ATOMICS lag
}
\details{
For details regarding the SAPPHIRE plot, please refer to the relative publications \url{http://www.nature.com/articles/srep06264}. 
Main documentation of the original campari software \url{http://campari.sourceforge.net/documentation.html}.
}
