% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_campari.R
\name{install_campari}
\alias{install_campari}
\title{Installing the original Fortran code}
\usage{
install_campari(installation_location = NULL, install_ncminer = FALSE,
  install_threads = FALSE, install_mpi = FALSE)
}
\arguments{
\item{installation_location}{It defaults to the package installation path (package/inst/campari) but this can be copied in another directory and installed from there.
NB: it will contain a source directory with its configure file and therefore must be executable.}

\item{install_ncminer}{If true the executable for netcdf and ascii (tsv, csv) file handling will be installed on the top of normal installation.}

\item{install_threads}{If this option is true and you have some multithreading fortran compiler (e.g. openmp) the campari_threads will be installed.}

\item{install_mpi}{If this option is true and you have some MPI fortran compiler the campari_mpi will be installed.}
}
\description{
\code{install_campari} is able to install the original campari Fortran code in the default directory (package/inst) or in a specified one. 
     Please remember that make and configure are two fundamental steps in this process.
}
\details{
For details, please refer to the main documentation of the original campari software \url{http://campari.sourceforge.net/documentation.html}.
}
