% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords_from_keyfile.R
\name{keywords_from_keyfile}
\alias{keywords_from_keyfile}
\title{Transform a keyfile in a list of keywords}
\usage{
keywords_from_keyfile(key_file_input, return_table = FALSE,
  keyword_list = NULL, return_string_of_arguments = FALSE,
  keyword_list_first = TRUE)
}
\arguments{
\item{key_file_input}{An already formatted keyfile. It can contain comments (#) and blank lines. Please note that in the case of multivariable keywords 
the algorithm will keep only the first value. Consider using \code{return_table} option in such a case.}

\item{return_table}{Defaults to \code{FALSE}. It will return a formatted table instead of a named list.}

\item{keyword_list}{If provided, the function will append the readed keywords to this list and return it.}

\item{return_string_of_arguments}{This argument will provide the possibility to have a string with already formatted function arguments (can be directly copied in \code{\link{run_campari}}).}

\item{keyword_list_first}{It defines if the inserted keyword_list must be inserted afterwords or before (\code{TRUE}) the loaded keyfile keywords.}
}
\value{
A named list or a table of keywords readed from the keyfile in input
}
\description{
Transform an input keyfile in a list (or table) of keywords which can be modified and used again in \code{\link{run_campari}}. 
     If \code{return_table} is \code{TRUE} a data.frame of the keywords is returned (it should be used mainly in the case of multivalues keywords, 
     that would be truncated to the first value otherwise).
}
\examples{
\dontrun{
 keywords_from_keyfile("keyfile.key")
}

}
\seealso{
\code{\link{run_campari}}, \code{\link{sapphire_plot}}.
}
