% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{load_trj_dcd}
\alias{load_trj_dcd}
\title{Load trajectory}
\usage{
load_trj_dcd(t_file)
}
\arguments{
\item{fl}{file location}

\item{n_fold}{Number of links contractions (folds)
Default: \code{0}}
}
\value{
tree: degree list, connectivity matrix and weights
}
\description{
This is a wonderful description(X)
}
\examples{
tree <- adjl_from_pi(fil = "data/PROGIDX_000000000001.dat")

}

