% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicate_trj.R
\name{multiplicate_trj}
\alias{multiplicate_trj}
\title{Multiplicate the trajectory variables}
\usage{
multiplicate_trj(trj, window = NULL, overlapping_reduction = NULL)
}
\arguments{
\item{trj}{Input trajectory (variables on the columns and equal-time spaced snpashots on the row). It must be a \code{matrix} or a \code{data.frame} of numeric.}

\item{window}{Number of snapshots taken before and after the considered snapshot.}

\item{overlapping_reduction}{}
}
\value{
It will return a modified trajectory matrix.
}
\description{
\code{multiplicate_trj} copy the future and past variables in the snapshot considered. This leads to a multiplication of the variables 
     for each snapshot, following the variable \code{window}.
}
\details{
#For more details, please refer to the main documentation of the original campari software \url{http://campari.sourceforge.net/documentation.html}.
}
\seealso{
\code{\link{adjl_from_progindex}}, \code{\link{gen_progindex}}, \code{\link{gen_annotation}}.
}
