% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_sapphire.R
\name{score_sapphire}
\alias{score_sapphire}
\title{Scoring the SAPPHIRE-plot basin division}
\usage{
score_sapphire(the_sap, ann, scoring_method = "nmi", silent = FALSE,
  plot_basin_identification = FALSE, nbinsxy = NULL,
  merge_clusters = FALSE, basin_optimization = NULL,
  number_of_clusters = NULL, force_matching = TRUE)
}
\arguments{
\item{the_sap}{Name of the PROGIDX_<...>.dat or REPIX_<...>.dat file or Data frame with three columns containing, in order, the (sorted) progress index,
the relative time indices and the cut function.}

\item{ann}{Annotation must be a single line with the true cluster labels.}

\item{scoring_method}{Precise scoring method (final comparison between cluster labels)
\itemize{
     \item "\code{mni}" 
     \item "\code{adjusted_rand_index}" 
     \item "\code{jaccard_index}" 
     \item "\code{purity}" 
}}

\item{silent}{A logical value indicating whether the function has to remain silent or not. Default value is \code{FALSE}.}

\item{plot_basin_identification}{A logical value indicating whether to display the SAPPHIRE plot with the resulting partitions or not. 
Black partitions are the matched ones, blue ones derive only from the dynamic analysis and orange ones only from the kinetic analysis. 
The green curve is the kinetic annotation (black curve) where the parabolic shape has been subtracted, i.e. the actual curve used for 
the peaks identification. Default value is \code{FALSE}.}

\item{nbinsxy}{Number of bins on the y and x axis of the 2-D histogram. Default to sqrt(nrow(the_sap)).}

\item{basin_optimization}{Method name for the selection of precise basin. This is a costly procedure which is based on an heuristical obj function w
with the following subgroups.
\itemize{
   \item "\code{uniformity}" not available
   \item "\code{number_of_clusters}" optimize on the basis of the number of clusters
   \item "\code{minimal_entropy}" not available yet
}}

\item{number_of_clusters}{if basin_optimization is active accordingly this must be set to integer.}

\item{force_matching}{Please refer to \code{\link{basin_recognition}} for further details about the match option.}

\item{merge_cluster}{Logical that allow clusters to be merged automatically if consecutives}
}
\value{
The resulting score (0-1)
}
\description{
\code{score_sapphire} uses the information provided by the SAPPHIRE basin recognition in order to score the identified free energy basins
    against the original partition (annotation).
}
\details{
For details regarding the SAPPHIRE plot, please refer to the relative publications \url{http://www.nature.com/articles/srep06264}. 
Main documentation of the original campari software \url{http://campari.sourceforge.net/documentation.html}.
}
\examples{

adjl <- mst_from_trj(trj = matrix(rnorm(1000), nrow = 100, ncol = 10))
ret<-gen_progindex(adjl = adjl)
gen_annotation(ret_data = ret, local_cut_width = 10)
\dontrun{
score_sapphire(the_sap = "PROGIDX_000000000001.dat", ann = rnorm(100))
}

}
