% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pdb.R
\name{write.pdb.d}
\alias{write.pdb.d}
\title{writing the pdb file}
\usage{
write.pdb.d(x, filename, round = FALSE, digit = 3)
}
\arguments{
\item{fl}{file location}

\item{n_fold}{Number of links contractions (folds)
Default: \code{0}}
}
\value{
tree: degree list, connectivity matrix and weights
}
\description{
pwriting the pdb file(X)
}

