% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zapphire.R
\name{zap_ggplot}
\alias{zap_ggplot}
\title{plotting}
\usage{
zap_ggplot(sap_file = NULL, sap_table = NULL, write = F,
  folderPlot = "plots/", timeline = T, subtitle = NULL, basin_call = F,
  local_cut = T, ann_trace = F, ann_trace_ret = F,
  background_height = NULL, ann_names_L = NULL, ann_names_R = NULL,
  title = "no title")
}
\arguments{
\item{fl}{file location}

\item{n_fold}{Number of links contractions (folds)
Default: \code{0}}
}
\value{
tree: degree list, connectivity matrix and weights
}
\description{
plotting from file(X)
}

