% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjl_from_adjmat.R
\name{adjl_from_adjmat}
\alias{adjl_from_adjmat}
\title{From adjacency matrix to adjacency list}
\usage{
adjl_from_adjmat(adj_m)
}
\arguments{
\item{adj_m}{Input matrix (adjacency matrix).}
}
\value{
A list of three elements: degree list, connectivity matrix and weights.
}
\description{
This function is able to transform a matrix of distances into an adjacency list that can be used for the analysis pipeline.
     Please remember that \code{gen_progindex} accepts only minimum spanning trees.
}
\seealso{
\code{\link{gen_progindex}}, \code{\link{mst_from_trj}}.
}
