% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{adjl_from_progindex}
\alias{adjl_from_progindex}
\title{Build the network from the already processed Progress Index file}
\usage{
adjl_from_progindex(prog_index_file)
}
\arguments{
\item{prog_index_file}{Progress index file location. This should be of the kind \code{"PROGIDX_000000000001.dat"} (original campari)
or \code{"REPIX_000000000001.dat"} (CampaRi).}
}
\value{
\code{adjl_from_progindex} will return a minimum spanning tree: degree list, connectivity matrix and weights
}
\description{
\code{adjl_from_progindex} is able to use the output file from original campari software
     (or the output file from \code{\link{gen_annotation}}) in order to generate again the minimum spanning tree.
}
\details{
For details, please refer to the main documentation of the original campari software \url{http://campari.sourceforge.net/documentation.html}
}
\examples{
\dontrun{
adjl <- adjl_from_progindex(fil = "PROGIDX_000000000001.dat")
}


}
\seealso{
\code{\link{gen_progindex}}, \code{\link{gen_annotation}}
}

