% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{adjl_from_trj}
\alias{adjl_from_trj}
\title{Build the network from the trajectory file}
\usage{
adjl_from_trj(trj, distance_method = 5, distance_weights = NULL,
  clu_radius = NULL, clu_hardcut = NULL, normalize_d = TRUE,
  birch_clu = FALSE, min_span_tree = TRUE, mode = "fortran",
  rootmax_rad = NULL, tree_height = NULL, n_search_attempts = NULL,
  cores = NULL, logging = FALSE)
}
\arguments{
\item{fl}{file location}

\item{n_fold}{Number of links contractions (folds)
Default: \code{0}}
}
\value{
tree: degree list, connectivity matrix and weights
}
\description{
This is a wonderful description(X)
}

