% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_annotation.R
\name{gen_annotation}
\alias{gen_annotation}
\title{Compute the annotation of the progrex index}
\usage{
gen_annotation(ret_data, local_cut_width = NULL, snap_start = NULL)
}
\arguments{
\item{ret_data}{This input must be the output from the function \code{gen_progindex}.}

\item{local_cut_width}{This variable defines the width for a 3 groups annotation of the progrex index.}

\item{snap_start}{Starting snapshot. It must be the same as in gen_progindex. Best to leave not assigned.}
}
\description{
\code{gen_annotation} is able to annotate the progrex index using the mean first passage time between groups of snapshots on the progrex index.
     This function will produce a file (e.g. \code{"REPIX_000000000001.dat"}) that contains all the information about progress index and annotations
     and it can be used to plot the sapphire plot using \code{sapphire_plot}.
}
\details{
For details, please refer to the main documentation of the original campari software \url{http://campari.sourceforge.net/documentation.html}.
}
\examples{
adjl <- mst_from_trj(trj = matrix(rnorm(1000), nrow = 100, ncol = 10))
adjl2 <- contract_mst(adjl = adjl)
ret<-gen_progindex(adjl = adjl2)
gen_annotation(ret_data = ret, local_cut_width = 10)


}
\seealso{
\code{\link{gen_progindex}}, \code{\link{sapphire_plot}}.
}

