% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_progindex.R
\name{gen_progindex}
\alias{gen_progindex}
\title{Compute the progress index along with parent and distance-to-parent information}
\usage{
gen_progindex(adjl = NULL, nsnaps = NULL, snap_start = 1)
}
\arguments{
\item{adjl}{A list of three elements: degree list, connectivity matrix and weights. This input must be a minimum spanning tree, otherwise
the function will not be able to resolve the progress index in a unique solution. If backend netcdf data handling is active, this input can be left
empty. In this case please set the general option \code{CampaRi.data_management="netcdf"}.}

\item{nsnaps}{Number of snashots in the trajectory. If \code{adjl} is inserted this specification is not necessary. Instead if the netcdf backend data handling
is active this value will be necessary and precise in order to avoid unpredictable behaviours.}

\item{snap_start}{The snapshot from which the progrex index will start.}
}
\value{
The output of this function MUST be used as input for the \code{\link{gen_annotation}} function.
}
\description{
This function is able to create the progrex index of a time series from its minimum spanning tree of the distances. 
     In order to understand how the progrex index is generated please refer to \url{http://www.sciencedirect.com/science/article/pii/S0010465513002038}.
}
\details{
For details, please refer to the main documentation of the original campari software \url{http://campari.sourceforge.net/documentation.html}.
}
\examples{
adjl <- mst_from_trj(trj = matrix(rnorm(1000), nrow = 100, ncol = 10))
ret <- gen_progindex(adjl = adjl)

}

