% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_network.R
\name{generate_network}
\alias{generate_network}
\title{Create a network for each snapshots}
\usage{
generate_network(trj, window = NULL, method = "wgcna", pp_method = NULL,
  overlapping_reduction = NULL, transpose_trj = FALSE, silent = FALSE,
  ...)
}
\arguments{
\item{trj}{Input trajectory (variables on the columns and equal-time spaced snpashots on the row). It must be a \code{matrix} or a \code{data.frame} of numeric.}

\item{window}{Number of snapshots taken before and after the considered snapshot.}

\item{method}{Supported pairwise similarity measures are 'wgcna', 'binary', 'euclidean', 'maximum', 'canberra', 'minkowski', 'covariance', 
'MI_MM'. If set 'none' the windows will be used only for the pp_method (works only with path_* or tsne/svd).}

\item{pp_method}{'path_euclidean', 'path_minkowski', 'path_manhattan', 'path_maximum', 'path_canberra', 'path_binary' will find the distance path to make
an hamiltonian cycle. 'svd' if a singular vector decomposition will be used to have the diagional of the d matrix as output. With 'SymmetricUncertainty' you will have
AdjacencySymmetricUncertainty=2*I(X;Y)/(H(X)+H(Y)) [method='MI_*' must be active]. 'tsne' will make a tsne transformation.}

\item{overlapping_reduction}{Not yet supported (It will if a great number of snapshots will be considered in this analysis - snapshot selection)}

\item{transpose_trj}{Defaults to F. If set T the junk of trj (i.e. with a specific window) is transposed so to infer a 
network with dimensions window*(window-1)/2}

\item{silent}{A logical value indicating whether the function has to remain silent or not. Default value is \code{FALSE}.}

\item{...}{Various variables. Possible values are \code{c('wgcna_type', 'wgcna_power', 'wgcna_corOp')}.}
}
\value{
It will return a modified trajectory matrix.
}
\description{
\code{generate_network} creates a sequence of network (one for each snapshot) using a sliding window of trajectory (see \code{window}). 
     It is also possible to reduce the dimensionality using an overlapping reduction (see \code{overlapping_reduction})
}
\details{
From WGCNA::adjacency: Correlation and distance are transformed as follows: for type = "unsigned", 
adjacency = |cor|^power; for type = "signed", adjacency = (0.5 * (1+cor) )^power; for type = "signed hybrid",
adjacency = cor^power if cor>0 and 0 otherwise; and for type = "distance", adjacency = (1-(dist/max(dist))^2)^power.
For more details on the SAPPHIRE plot, please refer to the main documentation of the original 
campari software \url{http://campari.sourceforge.net/documentation.html}.
}
\seealso{
\code{\link{adjl_from_progindex}}, \code{\link{gen_progindex}}, \code{\link{gen_annotation}}.
}
