% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nSBR.R
\name{nSBR}
\alias{nSBR}
\title{New identifying basins in the SAPPHIRE plot}
\usage{
nSBR(data, ny, local.cut = FALSE, comb_met = c("MIC"), n.cluster = NULL,
  force_correct_ncl = FALSE, unif.splits = NULL, pk_span = NULL,
  plot = FALSE, silent = FALSE, return_plot = FALSE,
  random_picks = NULL, ann = NULL, shuffles = FALSE, ...)
}
\arguments{
\item{data}{Name of the PROGIDX_<...>.dat file or Data frame with three columns containing, in order, the (sorted) progress index, the relative time 
indices and the cut function.}

\item{ny}{Number of bins on the y-axis of the histogram. Finer cuts can detect smaller differencies, but they could stretch to the meaningless when data points
are sparse.}

\item{local.cut}{Logical value indicating whether the localized cut function has to be used (see references). Default value is \code{FALSE}.}

\item{comb_met}{Method to weight the barriers. One of the following:
\itemize{
      \item "\code{MIC}" Use this to weight the barriers with the Maximal Information Coefficient.
      \item "\code{MIC_kin}" Combine the MIC score and the kinetic annotation for weighting the barriers.
      \item "\code{kin}" Use only kinetic annotation for weighting the barriers.
      \item "\code{diff_kin}" Better not to use.
      \item "\code{diff}" Better not to use.
    }}

\item{n.cluster}{Integer. If this value is inserted only the first \code{n.cluster - 1} barriers are shown with dotted lines on the plot. Note that no
preselection of the barriers is done in the final output and this must be done again in a second instance. The output contains all the possible barriers (ordered).}

\item{force_correct_ncl}{A logical value indicating whether the correct number of clusters (in a minimum sense) must be respected. In practice it returns the truncated
list with a considerable number of warnings. Default value is \code{FALSE}.}

\item{unif.splits}{Integer. This variable defines the number of splits for the uniform sampling. Using only one number, e.g. 40 divisions, 
could not be optimal for clogged data. Generally speaking, this value is useful for the MI - ratio weighting of the barriers. It is also possible to insert a 
vector of integer and the resulting value is an average of these expanded results. Standard is \code{c(5,10,15,20,25,30,40,50,60)} generally.}

\item{pk_span}{Integer. The spanning window that should be the minimum size of the basin. It defines the peak finding step.}

\item{plot}{A logical value indicating whether to display the SAPPHIRE plot with the resulting partitions or not. The dark red curve indicates the MIC interpolation
while black dots indicate the peaks found. Here it is possible to see the effect of the spanning variable. Grey dots in the background are the temporal annotation as
it has been ordere by the progress index.}

\item{silent}{A logical value indicating whether the function has to remain silent or not. Default value is \code{FALSE}.}

\item{return_plot}{Logical. This can return the plot object in the return list.}

\item{random_picks}{Insert the number of times the barriers should be randomly recalculated for having a random baseline.}

\item{ann}{If random_picks is inserted you need to score the the random barriers against an annotation!!}

\item{shuffles}{A logical value indicating whether the data must be shuffled before analysis (both data and ann). Default value is \code{FALSE}.}

\item{...}{\itemize{
    \item "\code{time.series}" File name. If specified, it substitutes the time series of the PROGIDX_<..> file with the one provided by the file.
}}
}
\value{
A list containing
\itemize{
        \item "\code{nbins}" c(ny, ny). This is not influential.
        \item "\code{barriers}" Found barriers (no cut is made with the number of clusters which is only for plotting).
        \item "\code{plot}" The plot if you wanted it returned.
        \item "\code{rnd.picks}" This is an object containing the results of the random.picks, i.e. a list contatining:
        \itemize{
                \item "\code{ggp.bar}" ggplot containing the barriers found and the scores found with the random procedure.
                \item "\code{ggp.dist.rnd.bar}"  ggplot containing the estimate (in z-scores) of the distance from random distribution (showed in light blue).
                \item "\code{rnd.scores.nas}" Number of NAs in the randomly picked scores.
                \item "\code{rnd.scores}" data.frame with all the scores found.
                \item "\code{max.rnd.scor}" score_sapphire obj on the best random pick.
                \item "\code{max.rnd.bar}" data.frame with the best parriers found randomly and their score (xm = barrier points, ym = score).
                \item "\code{ref.score}" score_sapphire obj with the standard best score optimization procedure.
                \item "\code{z.scores}" Found zscore between optimal barrier and random picks distribution.
                \item "\code{std.n.bar}" Total number of barriers found without imposing the number of clusters in optimization procedure.
                \item "\code{pval_zsc}" The p-value calculated using the 
                \item "\code{pval_95per}" Total number of barriers found without imposing the number of clusters in optimization procedure.
        }
      }
}
\description{
Extimation of the MI curve on the temporal annotation (dots). Using the inverted MIC in particular we can find the clearest separations between basins
    (barriers) by looking at the highest values. To calculate this line we randomly split the PI (x-axis) in randomly chosen uniformly sized pieces and we 
    calculate the y-axis histogram to compare using MIC. We invert it and we interpolate all the results.
}
\details{
For details regarding the SAPPHIRE plot, please refer to the relative publications \url{http://www.nature.com/articles/srep06264}. 
Main documentation of the original campari software \url{http://campari.sourceforge.net/documentation.html}.
}
