% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nSBR.R
\name{nSBR}
\alias{nSBR}
\title{New identifying basins in the SAPPHIRE plot}
\usage{
nSBR(data, ny, local.cut = FALSE, n.cluster = NULL, comb_met = c("MIC"),
  unif.splits = NULL, pk_span = NULL, plot = FALSE, silent = FALSE, ...)
}
\arguments{
\item{data}{Name of the PROGIDX_<...>.dat file or Data frame with three columns containing, in order, the (sorted) progress index, the relative time indices and the cut function.}

\item{ny}{Number of bins on the y-axis of the 2-D histogram. Default to nx.}

\item{local.cut}{Logical value indicating whether the localized cut function has to be used (see references). Default value is \code{FALSE}.}

\item{plot}{A logical value indicating whether to display the SAPPHIRE plot with the resulting partitions or not. Black partitions 
are the matched ones, blue ones derive only from the dynamic analysis and orange ones only from the kinetic analysis. The green curve is the kinetic annotation (black curve) 
where the parabolic shape has been subtracted, i.e. the actual curve used for the peaks identification. Default value is \code{FALSE}.}

\item{silent}{A logical value indicating whether the function has to remain silent or not. Default value is \code{FALSE}.}

\item{...}{\itemize{
    \item "\code{time.series}" File name. If specified, it substitutes the time series of the PROGIDX_<..> file with the one provided by the file.
    \item "\code{cl.stat}" If \code{TRUE} it will activate the cluster analysis with Mutual Information, Hellinger distance and Shannon Entropy
    as defaults.
    \item "\code{cs.nBreaks}" Integer. This variable defines the number of splits for the analysis. If set to 0 it will use the breaks found in the SBR.
    \item "\code{cs.denat}" This value can be set to \code{"process_subtraction"} or \code{"poly_interpolation"} and it is defining the removal of 
    parabolic artifacts in the kinetic trace. The polynomial fit is by default 7 and 12 in degree for the kinetic annotation and uniform MI curves. The process 
    option istead is referring to the simulated process which is behind the kin ann. This option can be set to \code{TRUE} for using the process way.
    \item "\code{cs.denat.MI}" Integer. If NULL its default is 7 for poly_interpolation of the kinetic ann but nothing is done on the MI score. 
    \item "\code{cs.MI_comb}" This value is the representative of how the various calculation in barrier weighting are combined. In particular the options 
    are the following:
    \itemize{
      \item "\code{mean}" Mean between the MI based on the SBR and the MI based on uniform subdivisions
      \item "\code{multip}" Multiplication between the MI based on the SBR and the MI based on uniform subdivisions
      \item "\code{kin_MI}" Mean between the kinetic annotaton and the MI based on the SBR and the MI based on uniform subdivisions
      \item "\code{kin}" Only kinetic annotation value on the barrier
      \item "\code{MI}" Only the MI based on the SBR and the MI based on uniform subdivisions
    }
    \item "\code{cs.nUni}" Integer. This variable defines the number of splits for the uniform sampling. If not set the algorithm will use 40 divisions which 
    could not be optimal for clogged data. Generally speaking, this value is useful for the MI - ratio weighting of the barriers. It is also possible to insert a 
    vector of integer and the resulting value is an average of these expanded results. We advice to use \code{c(5,10,15,20,25,30,40,50)} generally.
    \item "\code{cs.entropy}" Logical. Calculate Shannon entropy.
    \item "\code{cs.stft}" Logical. Calculate short time fourier transform.
    \item "\code{cs.TE}" Logical. Calculate symmetric Transfer Entropy.
    \item "\code{cs.KL}" Logical. Calculate symmetric and non-symmetric Kullback-Leibler divergence.
    \item "\code{cs.wMI}" Logical. Use the number of breaks (\code{cs.nBreaks}) to find the Mutual Information for 10 slided divisions.
    This will result in 10*nBreaks values.
    \item "\code{data.out.it}" Should I output the data.frame in input?
}}

\item{nx}{Number of bins on the x-axis of the 2-D histogram.}

\item{ny.aut}{Logical value indicating whether a suitable number of bins on the y-axis has to be identified automatically. Default value is \code{FALSE}.}
}
\value{
A list containing
      \itemize{
        \item "\code{tab.st}" Data frame containing the boundaries of each state, their lengths and the type of the right boundary. If the cl.stat option
        is active this will have other columns with the barrier statistics (e.g. Hellinger distance).
        Type=1 means that it is a matched partition, type=2 is only dynamic, type=3 is only kinetic. The type of the last state is always equal to 1.
        \item "\code{nbins}" 2-D vector containing number of bins on x-axis and y-axis.
        \item "\code{seq.st}" The time-ordered discretized trajectory.
        \item "\code{statistics}" If \code{cl.stat} is \code{TRUE} this element will contain all the cluster statistics (unbound to the found barriers). 
        Otherwise it is \code{NULL}. If no barrier has been found this value defaults to \code{FALSE}.
        \item "\code{call}" The matched call. 
        \item "\code{data}" The sapphire file name if it is provided, otherwise the data obj. It is used by \code{\link{score_sapphire}}.
      }
}
\description{
Todo.
}
\details{
For details regarding the SAPPHIRE plot, please refer to the relative publications \url{http://www.nature.com/articles/srep06264}. 
Main documentation of the original campari software \url{http://campari.sourceforge.net/documentation.html}.
}
\examples{

adjl <- mst_from_trj(trj = matrix(rnorm(1000), nrow = 100, ncol = 10))
ret<-gen_progindex(adjl = adjl)
gen_annotation(ret_data = ret, local_cut_width = 10)
\dontrun{
basins_recognition("PROGIDX_000000000001.dat", nx=500)
}

}
