% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zapphire.R
\name{sapphire_plot}
\alias{sapphire_plot}
\title{Sapphire plot - plotting annotation and progrex index}
\usage{
sapphire_plot(sap_file = NULL, sap_table = NULL, write = F,
  folderPlot = "plots/", return_plot = F, local_cut = TRUE,
  timeline = FALSE, sub_sampling_factor = NULL, ann_trace = FALSE,
  return_ann_trace = FALSE, title = "", ...)
}
\arguments{
\item{sap_file}{Annotated progrex index file. This can be left unused if \code{sap_table} is used instead.
This input must be the file output of \code{gen_annotation}.}

\item{sap_table}{Annotated progrex index table. As for \code{sap_file} it can be left unused if the other input is used.
Also in this case the format of the table must be the file output of \code{gen_annotation}.}

\item{write}{A logical indicating whether to write the plot to file.}

\item{folderPlot}{A character string that defines the folder in which save the plots (\code{write} argument must be \code{TRUE}).}

\item{return_plot}{if \code{TRUE} it returns the plot object (ggplot).}

\item{local_cut}{A logical that, if \code{TRUE} plots the local cut annotation function (see \code{gen_annotation}).}

\item{timeline}{A logical indicating whether add a timeline annotation (with the original order) on the bottom of the plot.}

\item{sub_sampling_factor}{if a number is inserted the annotation will be subsampled by that factor.}

\item{ann_trace}{This argument can be set to logical, an integer or a matrix of integers. If logical, the annotation will be a horizontal thick line
that follows a split in half of the original timeline. If only an integer have been inserted the time line will be split in that number of splits.
Instead if a matrix have been inserted it will be a plotted as it is on the top of the plot mapping the range of numbers in order to
produce a gray scale horizontal annotation.}

\item{return_ann_trace}{If \code{TRUE} the annotation vector is returned. This option can be really useful for adding layers and specific text \link{ggplot2}.}

\item{title}{Title of the plot (default "")}
}
\value{
If \code{ann_trace_ret} is active it will return the annotation trace used for the plot.
}
\description{
\code{sapphire_plot} is able to generate a sapphire plot using the progrex index file (e.g. \code{"REPIX_000000000001.dat"})
     generated by \code{\link{gen_annotation}}.
}
\details{
For details, please refer to the main documentation of the original campari software \url{http://campari.sourceforge.net/documentation.html}.
}
\examples{
adjl <- mst_from_trj(trj = matrix(rnorm(1000), nrow = 100, ncol = 10))
ret<-gen_progindex(adjl = adjl)
gen_annotation(ret_data = ret, local_cut_width = 10)
\dontrun{
zap_ggplot("REPIX_000000000001.dat")
}

}
\seealso{
\code{\link{mst_from_trj}}, \code{\link{gen_progindex}}, \code{\link{gen_annotation}}.
}
