% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_sapphire.R
\name{score_sapphire}
\alias{score_sapphire}
\title{Scoring the SAPPHIRE-plot basin division}
\usage{
score_sapphire(the_sap, ann, manual_barriers = NULL, basin_obj = NULL,
  scoring_method = "adjusted_rand_index", multi_cluster_policy = "popup",
  plot_pred_true_resume = FALSE, silent = FALSE, return_plot = FALSE,
  return_predicted = FALSE, ...)
}
\arguments{
\item{the_sap}{Name of the PROGIDX_<...>.dat or REPIX_<...>.dat file or Data frame with three columns containing, in order, the (sorted) progress index,
the relative time indices and the cut function.}

\item{ann}{Annotation must be a single line with the true cluster labels.}

\item{manual_barriers}{If an integer vector is inserted, it is used as the barrier locations.}

\item{basin_obj}{Output of \code{\link{basins_recognition}}. If you used \code{\link{basin_optimization}} to optimize the clusters please
insert only the resulting bas object}

\item{scoring_method}{Precise scoring method (final comparison between cluster labels)
\itemize{
     \item "\code{nmi}" 
     \item "\code{adjusted_rand_index}" 
     \item "\code{jaccard_index}" 
     \item "\code{purity}" 
}}

\item{multi_cluster_policy}{This string decides how the clusters must be handled in the case of more clusters found in comparison to the annotation 
inserted. The default is \code{'popup'}. The available values are:
     \itemize{
        \item "\code{popup}" Creates new clusters (ordered by shannon weight).
        \item "\code{keep}" Keeps the duplications and so on. This is a bias feature.
        \item "\code{merge_previous}" Takes in account the last label inserted in the shannon weigthed table.
      }}

\item{plot_pred_true_resume}{Defaults tp \code{FALSE}. If set to true it plots the predicted and true labels one along the other.}

\item{silent}{A logical value indicating whether the function has to remain silent or not. Default value is \code{FALSE}.}

\item{return_plot}{returns plots}

\item{return_predicted}{If true return the predicted vector.}

\item{...}{\code{max_number_of_elements} for plotting can be supplied. It defaults to 20k.}
}
\value{
A list containing
      \itemize{
        \item "\code{score.out}" Resulting score between 0 and 1 (good).
        \item "\code{label_freq_list}" Total representation of the clusters along with the single shannon entropies.
        \item "\code{main_desc}" Description of the main clusters found.
        \item "\code{miss_perc}" Percentage of not correctly defined divisions.
      }
}
\description{
\code{score_sapphire} uses the information provided by the SAPPHIRE basin recognition in order to score the identified free energy basins
    against the original partition (annotation).
}
\details{
For details regarding the SAPPHIRE plot, please refer to the relative publications \url{http://www.nature.com/articles/srep06264}. 
Main documentation of the original campari software \url{http://campari.sourceforge.net/documentation.html}.
}
\examples{
adjl <- mst_from_trj(trj = matrix(rnorm(1000), nrow = 100, ncol = 10))
ret<-gen_progindex(adjl = adjl)
gen_annotation(ret_data = ret, local_cut_width = 10)
\dontrun{
score_sapphire(the_sap = "PROGIDX_000000000001.dat", ann = rnorm(100))
}

}
