% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_sapphire.R
\name{score_sapphire}
\alias{score_sapphire}
\title{Scoring the SAPPHIRE-plot basin division}
\usage{
score_sapphire(the_sap, ann, manual_barriers = NULL, basin_obj = NULL,
  scoring_method = "nmi", merge_clusters = FALSE,
  plot_pred_true_resume = FALSE, silent = FALSE, ...)
}
\arguments{
\item{the_sap}{Name of the PROGIDX_<...>.dat or REPIX_<...>.dat file or Data frame with three columns containing, in order, the (sorted) progress index,
the relative time indices and the cut function.}

\item{ann}{Annotation must be a single line with the true cluster labels.}

\item{manual_barriers}{If an integer vector is inserted, it is used as the barrier locations.}

\item{basin_obj}{Output of \code{\link{basins_recognition}}. If you used \code{\link{basin_optimization}} to optimize the clusters please
insert only the resulting bas object}

\item{scoring_method}{Precise scoring method (final comparison between cluster labels)
\itemize{
     \item "\code{nmi}" 
     \item "\code{adjusted_rand_index}" 
     \item "\code{jaccard_index}" 
     \item "\code{purity}" 
}}

\item{merge_clusters}{Logical that allow clusters to be merged automatically if consecutives}

\item{plot_pred_true_resume}{Defaults tp \code{FALSE}. If set to true it plots the predicted and true labels one along the other.}

\item{silent}{A logical value indicating whether the function has to remain silent or not. Default value is \code{FALSE}.}

\item{...}{Not yet in use.}

\item{merge_clusters}{A logical value for indicating if merging consecutive clusters with same labels.}
}
\value{
A list containing
      \itemize{
        \item "\code{score.out}" Resulting score between 0 and 1 (good).
        \item "\code{max_freq_table}" Matrix with a description of the labels selected and their cluster characteristics
        \item "\code{label_freq_list}" Total representation of the clusters
      }
}
\description{
\code{score_sapphire} uses the information provided by the SAPPHIRE basin recognition in order to score the identified free energy basins
    against the original partition (annotation).
}
\details{
For details regarding the SAPPHIRE plot, please refer to the relative publications \url{http://www.nature.com/articles/srep06264}. 
Main documentation of the original campari software \url{http://campari.sourceforge.net/documentation.html}.
}
\examples{
adjl <- mst_from_trj(trj = matrix(rnorm(1000), nrow = 100, ncol = 10))
ret<-gen_progindex(adjl = adjl)
gen_annotation(ret_data = ret, local_cut_width = 10)
\dontrun{
score_sapphire(the_sap = "PROGIDX_000000000001.dat", ann = rnorm(100))
}

}
