% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_sapphire.R
\name{score_sapphire}
\alias{score_sapphire}
\title{Scoring the SAPPHIRE-plot basin division}
\usage{
score_sapphire(the_sap, ann, scoring_method = "nmi", merge_clusters = FALSE,
  basin_optimization = FALSE, plot_basin_identification = FALSE,
  plot_pred_true_resume = FALSE, nbins_x = NULL, nbins_y = nbins_x,
  force_matching = FALSE, silent = FALSE, ...)
}
\arguments{
\item{the_sap}{Name of the PROGIDX_<...>.dat or REPIX_<...>.dat file or Data frame with three columns containing, in order, the (sorted) progress index,
the relative time indices and the cut function.}

\item{ann}{Annotation must be a single line with the true cluster labels.}

\item{scoring_method}{Precise scoring method (final comparison between cluster labels)
\itemize{
     \item "\code{mni}" 
     \item "\code{adjusted_rand_index}" 
     \item "\code{jaccard_index}" 
     \item "\code{purity}" 
}}

\item{basin_optimization}{If \code{TRUE} it will use \code{\link{basin_optimization}} to optimize the clusters. Please consider adding its relevant variables.}

\item{plot_basin_identification}{A logical value indicating whether to display the SAPPHIRE plot with the resulting partitions or not. 
Black partitions are the matched ones, blue ones derive only from the dynamic analysis and orange ones only from the kinetic analysis. 
The green curve is the kinetic annotation (black curve) where the parabolic shape has been subtracted, i.e. the actual curve used for 
the peaks identification. Default value is \code{FALSE}.}

\item{plot_pred_true_resume}{Defaults tp \code{FALSE}. If set to true it plots the predicted and true labels one along the other.}

\item{nbins_x}{Number of bins on  x-axis of the 2-D histogram. Default to sqrt(nrow(the_sap)).}

\item{nbins_y}{Number of bins on the y-axis of the 2-D histogram. Default to sqrt(nrow(the_sap)).}

\item{force_matching}{Please refer to \code{\link{basin_recognition}} for further details about the match option.}

\item{silent}{A logical value indicating whether the function has to remain silent or not. Default value is \code{FALSE}.}

\item{...}{This variables will be sent to \code{\link{basin_optimization}} without any checking}

\item{merge_cluster}{Logical that allow clusters to be merged automatically if consecutives}
}
\value{
A list containing
      \itemize{
        \item "\code{score.out}" Resulting score between 0 and 1 (good).
        \item "\code{max_freq_table}" Matrix with a description of the labels selected and their cluster characteristics
        \item "\code{label_freq_list}" Total representation of the clusters
      }
}
\description{
\code{score_sapphire} uses the information provided by the SAPPHIRE basin recognition in order to score the identified free energy basins
    against the original partition (annotation).
}
\details{
For details regarding the SAPPHIRE plot, please refer to the relative publications \url{http://www.nature.com/articles/srep06264}. 
Main documentation of the original campari software \url{http://campari.sourceforge.net/documentation.html}.
}
\examples{
adjl <- mst_from_trj(trj = matrix(rnorm(1000), nrow = 100, ncol = 10))
ret<-gen_progindex(adjl = adjl)
gen_annotation(ret_data = ret, local_cut_width = 10)
\dontrun{
score_sapphire(the_sap = "PROGIDX_000000000001.dat", ann = rnorm(100))
}

}
