% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{select_features}
\alias{select_features}
\title{Multiple pre-processing methods for feature selection}
\usage{
select_features(trj, feature_selection = "pca",
  n_princ_comp = floor(ncol(trj)/10), pca_method = "R", plotit = FALSE,
  frameit = FALSE, return_plot = FALSE, cluster_vector = NULL,
  plotly_it = FALSE, points_size = 1, specific_palette = NULL,
  plot_legend = FALSE, legend_title = NULL, legend_labels = NULL)
}
\arguments{
\item{trj}{Input trajectory (variables on the columns and equal-time spaced snpashots on the row). It must be a \code{matrix} or a \code{data.frame} of numeric.}

\item{feature_selection}{Available method is 'pca'}

\item{n_princ_comp}{number of principal components to use}

\item{pca_method}{If set 'R' (default) it will use \code{\link{princomp}}. The other (slower) option is 'robust' which is using \code{\link{PCAproj}}.}

\item{plotit}{Plot the PCA components if two are selected.}

\item{frameit}{Add a frame (shaded clustering) of the whole performed PCA.}

\item{return_plot}{This option is usually used to add layers to the ggplot (made using autoplot).}

\item{cluster_vector}{This option can be used to set the clusters you want and show them with different colors (and shades if \code{frameit = TRUE}). 
Please set this option with the same dimensionality of the trj (n_snapshots) and use integer numbers (to define the clusters).}

\item{points_size}{It must be a number and it defines the size of the points.}
}
\value{
It will return a modified trajectory matrix and print the principal components vector.
}
\description{
\code{select_features} is able to select input variables on the basis of the trajectory input. For the moment only PCA-based feature selection
     is supported. Moreover, this tool is meant to be used with the total trajectory input.
}
\details{
This function is based primarly on the basic R function \code{pricomp} and on \code{PCAproj} from the package pcaPP. Insead, for more details on the SAPPHIRE anlysis, please refer to the main documentation
of the original campari software \url{http://campari.sourceforge.net/documentation.html}.
}
\seealso{
\code{\link{princomp}}, \code{\link{PCAproj}}, \code{\link{adjl_from_progindex}}, \code{\link{gen_progindex}}, \code{\link{gen_annotation}}.
}
