% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_clustering_summary.R
\name{show_clustering_summary}
\alias{show_clustering_summary}
\title{Shows the clustering summary from log file}
\usage{
show_clustering_summary(log_file, fyc_file, verbose = TRUE,
  number_of_clusters = 10, return_centers = FALSE, return_angles = FALSE)
}
\arguments{
\item{log_file}{log output of a campari run in which a clustering was performed}

\item{fyc_file}{the fyc file which was output of the same campari run}

\item{verbose}{if verbose is \code{TRUE} the clustering summary will be printed directly in the console (first 10 elements).}
}
\value{
A named list or a table of keywords readed from the keyfile in input
}
\description{
Using sed and awk to print a formatted cluster summary. log file and FYC file are needed.
}
\examples{
\dontrun{
 keywords_from_keyfile("keyfile.key")
}


}
\seealso{
\code{\link{run_campari}}, \code{\link{sapphire_plot}}.
}
