% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_clustering_summary.R
\name{show_clustering_summary}
\alias{show_clustering_summary}
\title{Shows the clustering summary from log file}
\usage{
show_clustering_summary(log_file, fyc_file = NULL, verbose = TRUE,
  which_first_clusters = NULL, which_cluster = NULL,
  sub_sampling_factor = 1, return_centers = FALSE, return_angles = FALSE)
}
\arguments{
\item{log_file}{log output of a campari run in which a clustering was performed}

\item{fyc_file}{the fyc file which was output of the same campari run. This is set only to work with the nbutane example (tutorial 11)}

\item{verbose}{if verbose is \code{TRUE} the clustering summary will be printed directly in the console (first 10 elements).}

\item{which_first_clusters}{It defaults to 10 and defines the number of centers to consider in the clustering summary.}

\item{which_cluster}{you can select a specific cluster element with this option}

\item{return_centers}{it returns the cluster centers. (fyc option)}

\item{return_angles}{it returns the cluster centers' angles. (fyc option)}
}
\value{
see return_*
}
\description{
Using sed and awk to print a formatted cluster summary. log file and FYC file are needed.
}
\examples{
\dontrun{
 keywords_from_keyfile("keyfile.key")
}


}
\seealso{
\code{\link{run_campari}}, \code{\link{sapphire_plot}}.
}
