% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pdb.R
\name{write.pdb.d}
\alias{write.pdb.d}
\title{writing the pdb file}
\usage{
write.pdb.d(x, base_name, round = FALSE, digit = 4)
}
\arguments{
\item{fl}{file location}

\item{n_fold}{Number of links contractions (folds)
Default: \code{0}}
}
\value{
tree: degree list, connectivity matrix and weights
}
\description{
pwriting the pdb file(X)
}
\details{
For details, please refer to the main documentation of the original campari software \url{http://campari.sourceforge.net/documentation.html}
}

