% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pdb.R
\name{write.pdb.d}
\alias{write.pdb.d}
\title{Writing a pdb file from trajectory matrix in R}
\usage{
write.pdb.d(trj, base_name, round = FALSE, digit = 4, dim_check = TRUE)
}
\arguments{
\item{trj}{Time series in a matrix shape (also data.frame numeric). The number of variables (nrow) must be divisible by 3 for atom-like pdb writing.}

\item{base_name}{File base name.}

\item{round}{\code{TRUE} will truncate the output following the \code{digit} variable.}

\item{digit}{Number of digits that will be kept from truncation (\code{round}).}
}
\description{
\code{write.pdb.d} writes to file a pdb file with the trajectory matrix in input. 
     The number of variables (number of columns) must be divisible by 3 to mantain the atomic 3D structure.
}
\details{
For details, please refer to the main documentation of the original campari software \url{http://campari.sourceforge.net/documentation.html}.
}
\examples{
write.pdb.d(matrix(rnorm(900), nrow = 100, ncol = 9), base_name = "rnorm_trj")
}
\seealso{
\code{\link{campari}}
}
