% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zapphire.R
\name{zap_ggplot}
\alias{zap_ggplot}
\title{samsahara}
\usage{
zap_ggplot(sap_file, write = F, folderPlot = "plots/", timeline = T,
  basin_call = F, ann_trace = NULL, ann_trace_ret = F,
  title = "no title")
}
\arguments{
\item{fl}{file location}

\item{n_fold}{Number of links contractions (folds)
Default: \code{0}}
}
\value{
tree: degree list, connectivity matrix and weights
}
\description{
plotting from file(X)
}
\examples{
none

}

